pyproject.toml
[project]
name = "receipe-transcriber"
version = "0.1.0"
description = ""
authors = [
    {name = "Tom Hanks",email = "tommy.hanks@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "flask-sqlalchemy>=3.0.0",
    "flask-migrate>=4.0.0",
    "flask-sse>=1.0.0",
    "python-dotenv>=1.0.0",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "ollama>=0.6.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "requests>=2.31.0",
    "gunicorn>=21.0.0",
    "gevent>=23.0.0",
]

[tool.poetry]
packages = [{include = "receipe_transcriber", from = "src"}]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m not slow')",
    "integration: marks tests as integration tests",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
